Role assignment and removal API
