--- dataModel/types/user.ts
@@ -1,9 +1,0 @@
-export interface User {
-    id: string;
-    email: string;
-    firstName: string;
-    lastName: string | null;
-    active: boolean;
-    notificationLevel: 'all' | 'necessary';
-    phone: string | null;
-}
