+++ dataStore/types/authenticationData.ts
