+++ routes/roles.ts
@@ -1,34 +1,24 @@
 import express from 'express';
 import {
-    assignRole,
-    changeRoleLevel,
-    deleteRole,
-    newRole
+    assignRole, removeRole,
 } from '../controllers/rolesController';
 import passport from 'passport';
+import { checkRole } from '../utils/roleAuthorization';
 
 export const roles = (app: express.Application): void => {
     const rolesRoute = express.Router();
     rolesRoute.get('/ping', (req, res) => res.send('pong').status(200).end());
     rolesRoute.post(
-        '/new',
-        passport.authenticate('jwt', { session: false }),
-        newRole
-    );
-    rolesRoute.post(
         '/assign',
         passport.authenticate('jwt', { session: false }),
+        checkRole(['unitTestAdmin', 'admin']),
         assignRole
     );
     rolesRoute.post(
-        '/delete',
+        '/remove',
         passport.authenticate('jwt', { session: false }),
-        deleteRole
-    );
-    rolesRoute.post(
-        '/changeLevel',
-        passport.authenticate('jwt', { session: false }),
-        changeRoleLevel
+        checkRole(['unitTestAdmin', 'admin']),
+        removeRole
     );
     app.use('/roles', rolesRoute);
 };
