+++ server/index.ts
@@ -11,8 +11,7 @@ import { authentication } from '../routes/auth';
 import { token } from '../routes/tokenRefresh';
 import { userdata } from '../routes/userData';
 import { verification } from '../routes/verification';
-import {roles} from "../routes/roles";
-import {tasks} from "../routes/tasks";
+import { roles } from '../routes/roles';
 
 const app = express();
 
@@ -28,7 +27,7 @@ token(app);
 userdata(app);
 verification(app);
 roles(app);
-tasks(app);
+
 
 app.use((req, res, next) => {
     const err = new Error('Not found');
