+++ tests/unit/rest/auth/passwordReset.test.ts
@@ -1,9 +1,4 @@
-import {
-    addUser,
-    deleteUser,
-    findUser,
-    getRefreshTokenFromAuthId,
-} from '../../../../services/dataStore';
+import { addUser } from '../../../../dataStore/types/user';
 
 process.env.NODE_ENV = 'test';
 import chai from 'chai';
@@ -11,6 +6,7 @@ import chaiHttp from 'chai-http';
 chai.use(chaiHttp);
 import { server } from '../../../../server';
 import { invalidUserEmail, validUser } from '../../../userConfiguration';
+import { deleteUser } from '../../../../dataStore/testingQueries';
 
 describe('Checks if the auth route is available', function () {
     it('Checks if the auth route is setup properly.', function (done) {
