+++ tests/unit/rest/userData/deleteUser.test.ts
@@ -1,8 +1,9 @@
-import { addUser, deleteUser, findUser } from '../../../../services/dataStore';
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import { server } from '../../../../server';
 import { invalidUserEmail, validUser } from '../../../userConfiguration';
+import { addUser, findUser } from '../../../../dataStore/types/user';
+import { deleteUser } from '../../../../dataStore/testingQueries';
 
 process.env.NODE_ENV = 'test';
 chai.use(chaiHttp);
