--- utils/roles.ts
@@ -1,5 +1,0 @@
-export const ROLES = {
-    Admin: 'Admin',
-    Pupil: 'Pupil',
-    Volunteer: 'Volunteer',
-};
